Split functions into different modules. Updated main function.
