+++ .gitignore
@@ -31,3 +31,6 @@
 # Debug files
 *.dSYM/
 *.su
+.idea/
+CMakeLists.txt
+cmake-build-debug/
