+++ filefuncs.h
@@ -1,0 +1,15 @@
+/*
+ * File Functions declaration - filefuncs.h
+ *
+ */
+
+#ifndef WHITEOS_PA1_FILEFUNCS_H
+#define WHITEOS_PA1_FILEFUNCS_H
+
+#include <stdio.h>
+#include <ctype.h>
+
+long int getStartPos(FILE *f);
+long int getEndPos(FILE *f);
+
+#endif //WHITEOS_PA1_FILEFUNCS_H
